aicap/requirements.txt
torch
torchvision
torchaudio
transformers==4.49.0
accelerate
pillow
typer[all]
rich
einops>=0.7.0
timm>=0.9.0